added some error handling to BanService and ReportService modules

+++ backend/modules/ban_module/BanService.js
@@ -2,13 +2,19 @@ const ERROR_CODES = require("./../../ErrorCodes.js")
 
 class BanService {
     async banUser(userID, userStore) {
+        if (!mongoose.isObjectIdOrHexString(userID)) {
+            return new ResponseObject(ERROR_CODES.INVALID)
+        }
         let response = await userStore.deleteUser(userID)
-        return new ResponseObject(ERROR_CODES.SUCCESS, response)
+        return response
     }
 
     async banEvent(eventID, eventStore) {
+        if (!mongoose.isObjectIdOrHexString(eventID)) {
+            return new ResponseObject(ERROR_CODES.INVALID)
+        }
         let response = await eventStore.deleteEvent(eventID)
-        return new ResponseObject(ERROR_CODES.SUCCESS, response)
+        return response
     }
 }
